refactor(#22): RankRepository dip 적용
